Day19 Solution
I needed a hint to get to the CYK algorithm.

Kept trying to build the possibilities for the strings and kept getting stuck.

After solving it, I looked at the other Advent solutions and saw a couple that used Regex, so I tried my hand at that one too. It's significantly faster... I wish I had gone with it to start. Learned for next year!